requirements.txt
Flask 
boto3
python-dotenv
psycopg2-binary